+++ bindings/lua/samples/modules/scripts/utils/utils.lua
@@ -1,0 +1,9 @@
+print("Utils loaded!")
+
+local utils = {}
+
+function utils.someutils()
+    print("Utils someutils called")
+end
+
+return utils